Fix macOS/arm64 startup + skirmish crashes; stop tracking buildMingw64
Get the native macOS (arm64) build booting to the title screen and into the
menus. Cross-platform bugs fixed:

- MachLogRaces ctor trap at -O2: sibling enums narrowed to `: unsigned char`
  were still iterated with `++((int&)x)` (a 4-byte read-modify-write through
  an int& onto a 1-byte object). Replaced all 64 `((int&)x)` sites across 20
  files with the width-safe `x = static_cast<decltype(x)>(x +/- 1)`; enum
  layout unchanged so persistence/network stay byte-identical.
- Malformed asset paths (filedata.cpp): SysFileData appended the
  MACH_ROOT-resolved leaf, yielding `<dir>//<root>/<leaf>`. Now appends the
  raw relative name via combine().
- SysPathName assert-logging infinite recursion -> stack overflow
  (pathname.cpp isAbsolute/isRelative PRE_INFO(*this)); log raw pathname_.
- Out-of-bounds skirmishSystem(CUSTOM): database.cpp always creates the CUSTOM
  system; databasi.cpp reserve 3 -> 4.
- DOS 8.3 validFileName rejected long paths like "machines-game-data"
  (prepost.cpp); relaxed to a non-empty check.
- Netcode remediation: message/random/netinet/netnet determinism and layout
  fixes toward cross-platform multiplayer.

Tooling: top-level Makefile (dev/run/debug/crash), scripts/crashrun.sh (lldb
crash catcher with backtrace dump), knowledge/ notes. Env-gated debug hook
MACH_AUTO_SKIRMISH (off by default).

Stop tracking buildMingw64/ (1199 generated files; already in .gitignore).

+++ include/libdev/mathex/random.hpp
@@ -51,8 +51,13 @@ private:
     //  to seedFromTime will use a different seed.
     static  size_t  seedIncrement();
 
-    ulong state_;
-    ulong seed_;
+    //  Fixed-width (uint32 == unsigned int, 32-bit on both LP64/macOS and LLP64/Windows).
+    //  Previously ulong, which is 64-bit on macOS but 32-bit on Windows, so the LCG below
+    //  (state_ = state_ * 1103515245 + 12345) wrapped at different widths and produced a
+    //  different sequence per platform - a determinism/desync source. uint32 wraps mod 2^32
+    //  identically everywhere.
+    uint32 state_;
+    uint32 seed_;
 };
 
 //  Acccessor functions for getting the results of a
